The range() method of LongStreamCompat now correctly accepts long start and end parameters.

+++ StreamCompat/src/main/java/com/github/wrdlbrnft/streamcompat/longstream/LongStreamCompat.java
@@ -13,7 +13,7 @@ public class LongStreamCompat {
         return new LongStreamImpl(iterator);
     }
 
-    public static LongStream range(int start, int end) {
+    public static LongStream range(long start, long end) {
         if (end < start) {
             throw new IllegalArgumentException("End value must be bigger than start value!");
         }
